+++ changes.md
@@ -17,5 +17,6 @@
 - Prefer direct comparison on github.com.
 - Support update arrow style in commit message subject (GNOME workflow only).
 - Support regenerate commits (Cinnamon workflow only).
+- Generate Xfce GitLab diff URL if the commit is likely related to Xfce updates.
 - Convert Xfce GitLab URL to GitHub mirror URL.
 - Sidebar width is now hardcoded. For now responsive is not a focus in this fork.